# openapi.yaml
openapi: 3.1.0
info:
  title: QuizForge API
  version: 0.5.0-dev
  summary: Asynchronous quiz and assessment tournaments.
  description: |
    QuizForge runs asynchronous quiz tournaments. A tournament opens for a
    window, players enter on their own schedule, and results roll up to a
    leaderboard. There is no synchronous "everyone answers at once" mode, by
    design.

    ## Status of this document

    This describes the API **exactly as it behaves today**.

    The shapes it first described were not all ones worth keeping. Six were
    corrected before anything consumed this document — while doing so was one
    commit rather than a breaking change — and are recorded in
    `docs/api/known-inconsistencies.md` with what they were and why they moved.

    ## Conventions

    - Every identifier is a prefixed `TypeId`: `acc_`, `wsp_`, `trn_`, `att_`,
      `qst_`. Bare UUIDs never appear in a response.
    - Every list endpoint returns `{ "data": [...], "nextCursor": ... }` and
      takes `limit` and `cursor`. Pagination is **keyset**, never offset: a
      cursor names the last row you saw, so a row inserted while you page
      through cannot duplicate or hide another one.
    - Out-of-range parameters are **rejected**, never silently corrected.

    Everything the product does is reachable here: workspaces, members, API
    keys, question banks, questions, tournaments and play.

    ## Questions are immutable

...
    question.

...

...

...

...
    have selected.

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - sessionCookie: []
  - bearerApiKey: []
paths:
...
  /v1/tournaments:
    post:
      tags: [Tournaments]
      operationId: createTournament
      summary: Schedule a tournament
      description: |
        Requires `EDITOR` or above.

        **A tournament that cannot be played is refused at creation**, not
        discovered by the first player: if the bank holds fewer current
        questions than `questions` asks for, this is a `400` naming both
        numbers. Retiring questions afterwards can still leave a scheduled
        tournament short, which is why the same check runs again on amendment.

        `opensAt` and `closesAt` define the window. State is derived from them
        and the clock, so a tournament is never "started" by anything.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TournamentDraft'
      responses:
        '201':
          description: The tournament was scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentSummary'
        '400':
          description: |
            The window is invalid, a count is out of range, or the bank holds
            too few questions. Carries `code: INVALID_REQUEST`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          description: |
            No such question bank in this workspace. Carries `code: NOT_FOUND` —
            a bank in another workspace is indistinguishable from one that does
            not exist.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          $ref: '#/components/responses/Internal'

    get:
      tags: [Tournaments]
      operationId: listTournaments
      summary: List tournaments in the selected workspace
      description: |
        Every tournament in the workspace, in no guaranteed order, with no
        pagination — the whole collection is returned.

        `state` is derived from the server clock at the moment of the request
        rather than stored, so a tournament transitions from `SCHEDULED` to
        `OPEN` to `CLOSED` without any job running, and a row can never
        disagree with the calendar.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: The workspace's tournaments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentPage'
        '400':
          $ref: '#/components/responses/InvalidListRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/tournaments/{tournamentId}:
    get:
      tags: [Tournaments]
      operationId: getTournament
      summary: One tournament
      description: |
        Its state is computed from the clock at the moment you read it, so the
        same tournament can answer `SCHEDULED` and then `OPEN` without anything
        having written to it.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - name: tournamentId
          in: path
          required: true
          description: The tournament.
          schema:
            $ref: '#/components/schemas/TournamentId'
      responses:
        '200':
          description: The tournament.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentSummary'
        '400':
          $ref: '#/components/responses/WorkspaceRequired'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'

    patch:
      tags: [Tournaments]
      operationId: updateTournament
      summary: Amend a tournament that has not opened
      description: |
        **Refused once the tournament is `OPEN` or `CLOSED`**, with
        `code: INVALID_REQUEST` and a detail naming the state. An open
        tournament has players who started under its current rules; a closed one
        has results those rules explain. Editing either makes the record
        dishonest.

        That includes the name, which is only cosmetic — the rule is one rule
        rather than a per-field matrix, because a per-field matrix is what
        nobody remembers correctly.

        The whole draft is replaced and revalidated, including the bank's
        question count: retirements between scheduling and amending can leave a
        tournament asking for more questions than remain.

        Requires `EDITOR` or above.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - name: tournamentId
          in: path
          required: true
          description: The tournament to amend.
          schema:
            $ref: '#/components/schemas/TournamentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TournamentDraft'
      responses:
        '200':
          description: The amended tournament.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentSummary'
        '400':
          description: |
            The tournament has already opened, the window is invalid, or the
            bank no longer holds enough questions. Carries
            `code: INVALID_REQUEST`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'

    delete:
      tags: [Tournaments]
      operationId: deleteTournament
      summary: Delete a tournament nobody has played
      description: |
        A real delete, and deliberately a narrow one: only while `SCHEDULED`,
        and only with **zero attempts**.

        A tournament with attempts is somebody's record. Removing it would leave
        scores and standings that nothing explains, which is why banks and
        questions are archived and retired rather than deleted. A scheduled
        tournament with no attempts is just a plan, and deleting a plan loses
        nothing — so this one is a real delete rather than another tombstone.

        Attempts in **any** state count, including expired and abandoned ones.
        Someone whose attempt ran out of time still played it.

        Requires `EDITOR` or above.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - name: tournamentId
          in: path
          required: true
          description: The tournament to delete.
          schema:
            $ref: '#/components/schemas/TournamentId'
      responses:
        '204':
          description: The tournament is gone.
        '400':
          description: |
            The tournament has opened, or has attempts against it. Carries
            `code: INVALID_REQUEST` and a detail saying which.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qf_session
      description: |
        Opaque session token issued by `POST /v1/auth/login`, stored only as a
        SHA-256 digest on the server. Mutating requests authenticated this way
        are CSRF-protected and additionally require the `X-XSRF-TOKEN` header.
    bearerApiKey:
      type: http
      scheme: bearer
      description: |
        A workspace API key, sent as `Authorization: Bearer qf_…`. Shown once
        when created and stored only as a SHA-256 digest thereafter.

        Authenticates a workspace rather than a person, so endpoints that
        attribute an action to an account — starting and submitting attempts —
        reject it.
  parameters:
    WorkspaceHeader:
      name: X-QuizForge-Workspace
      in: header
      required: false
      description: |
        The workspace to operate on, as a `wsp_…` identifier.

        Required for every endpoint outside `/v1/auth` when authenticating with
        a session cookie; omitting it fails with `400 INVALID_REQUEST`. With an
        API key the workspace comes from the key and this header is ignored.

        Documented as `required: false` because it is genuinely optional for
        key-authenticated callers, not because it is safe to leave out.

        Naming a workspace you are not a member of returns
        **`403 PERMISSION_DENIED`**: the credential is valid, the workspace is
        not yours. A malformed value is `400 INVALID_REQUEST`.
      schema:
        $ref: '#/components/schemas/WorkspaceId'
    Limit:
      name: limit
      in: query
      required: false
      description: |
        How many items to return, 1–100.

        A value outside that range is `400 INVALID_REQUEST`. It is deliberately
        not clamped: answering a request for 1000 with 100 items is
        indistinguishable from a collection that only holds 100, so a client
        would conclude it had everything when it had a prefix.
      schema:
        type: integer
        format: int32
        default: 25
        minimum: 1
        maximum: 100
    PageCursor:
      name: cursor
      in: query
      required: false
      description: |
        The `nextCursor` from the previous page. Omit for the first page.

        **Treat this as opaque.** It is not a page number, an offset, or an
        identifier, and its encoding is not part of the contract. A value this
        API did not issue is `400 INVALID_CURSOR` rather than being quietly
        treated as "start again".
      schema:
        $ref: '#/components/schemas/CursorParameter'
...
  responses:
...
    WorkspaceRequired:
      description: |
        No workspace is in scope. Carries `code: INVALID_REQUEST`.

        Sent when a session-authenticated request omits
        `X-QuizForge-Workspace`. Tenant isolation depends on a workspace being
        selected, so such a request is refused before it reaches a handler
        rather than being served unfiltered.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    WorkspaceForbidden:
      description: |
        The credential is valid but the account is not a member of the workspace
        it named. Carries `code: PERMISSION_DENIED`.

        Distinct from `401` deliberately: re-authenticating cannot fix this, so
        telling the caller they are unauthenticated would send them to the wrong
        remedy.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationRequired:
      description: |
        No valid credential was presented. Carries
        `code: AUTHENTICATION_REQUIRED`.

        Also returned in two cases that are not obviously authentication
        failures:

        - A path that does not exist. Authentication is evaluated before
          routing, so an anonymous caller learns nothing about which paths are
          real.
        It is **not** returned for a workspace you are not a member of — that
        is `403`, because the credential itself is valid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InvalidListRequest:
      description: |
        The request could not be paged. One of:

        - no workspace is in scope — `code: INVALID_REQUEST`
        - `limit` is outside 1–100 — `code: INVALID_REQUEST`
        - `cursor` was not issued by this API — `code: INVALID_CURSOR`

        The three are distinguished by `code`, which is why they share a status
        without sharing a meaning.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: |
        No such object, **or** it belongs to another workspace or another
        account. The two are deliberately indistinguishable, so this endpoint
        cannot be used to discover that an identifier exists.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Internal:
      description: |
        Something failed unexpectedly. The cause is logged server-side and
        never included in the response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      title: Problem
      description: |
        An RFC 9457 Problem Details document. Every error in the API has this
        shape.

        Clients should branch on `code`. It is the only field with a stability
        guarantee: a code that exists will never change meaning or HTTP status,
        and new codes are added rather than existing ones repurposed.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          maxLength: 200
          description: Stable URI identifying the error class.
          examples: ['https://quizforge.dev/errors/not-found']
        title:
          type: string
          maxLength: 100
          description: The error code, repeated as a human-readable title.
          examples: ['NOT_FOUND']
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code, repeated in the body.
          examples: [404]
        detail:
          type: string
          maxLength: 2000
          description: |
            Human-readable explanation. **Not stable** — reworded freely. Never
            parse it.
          examples: ['no such attempt']
        instance:
          type: string
          format: uri-reference
          maxLength: 2000
          description: The path that produced the error.
          examples: ['/v1/attempts/att_01a0136fb3937209a5719e8fb7db26bf']
        code:
          $ref: '#/components/schemas/ErrorCode'
    ErrorCode:
      type: string
      title: ErrorCode
      description: |
        The machine-readable error code. Exhaustive: every error the API can
        produce carries one of these.

        Each maps to exactly one HTTP status, and that mapping is part of the
        contract.
      enum:
        - INVALID_REQUEST
        - INVALID_CREDENTIALS
        - AUTHENTICATION_REQUIRED
        - PERMISSION_DENIED
        - NOT_FOUND
        - ALREADY_EXISTS
        - RATE_LIMITED
        - INTERNAL
      examples: ['NOT_FOUND']
      x-error-codes:
        INVALID_REQUEST:
          status: 400
          meaning: Malformed input, a failed validation rule, or an operation the current state does not allow.
        INVALID_CREDENTIALS:
          status: 401
          meaning: The email or password was wrong. Which one is never disclosed.
        AUTHENTICATION_REQUIRED:
          status: 401
          meaning: No credential was presented, or it has expired or been revoked.
        PERMISSION_DENIED:
          status: 403
          meaning: Authenticated, but the role is insufficient for this action.
        NOT_FOUND:
          status: 404
          meaning: No such object, or it belongs to another tenant. Deliberately ambiguous.
        ALREADY_EXISTS:
          status: 409
          meaning: A uniqueness constraint would be violated.
        RATE_LIMITED:
          status: 429
          meaning: Too many requests. Not yet emitted — rate limiting arrives with M4.
        INTERNAL:
          status: 500
          meaning: An unhandled failure. The cause is logged and never returned.
...
    WorkspaceId:
      type: string
      title: WorkspaceId
      description: A workspace identifier. UUIDv7 rendered as 32 lowercase hex characters with a `wsp_` prefix.
      pattern: '^wsp_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['wsp_01a0136fb3937209a5719e8fb7db26bf']
    TournamentId:
      type: string
      title: TournamentId
      description: A tournament identifier. UUIDv7 rendered as 32 lowercase hex characters with a `trn_` prefix.
      pattern: '^trn_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['trn_01a0136fb3937209a5719e8fb7db26bf']
...
    Cursor:
      type: [string, 'null']
      title: Cursor
      description: |
        Pass this back as `cursor` to get the next page. `null` means there is
        no next page — stop, rather than asking for one you already know is
        empty.

        **Opaque.** Not a page number, not an offset, not an identifier. Its
        encoding is not part of the contract and may change; a cursor issued by
        an older version will be rejected rather than misread.
      maxLength: 500
      examples: ['djE6MDFhMDEzNmZiMzkzNzIwOWE1NzE5ZThmYjdkYjI2YmY']
    CursorParameter:
      type: string
      title: CursorParameter
      description: An opaque cursor previously returned as `nextCursor`.
      maxLength: 500
      examples: ['djE6MDFhMDEzNmZiMzkzNzIwOWE1NzE5ZThmYjdkYjI2YmY']
    TournamentPage:
      type: object
      title: TournamentPage
      required: [data]
      properties:
        data:
          type: array
          maxItems: 10000
          items:
            $ref: '#/components/schemas/TournamentSummary'
        nextCursor:
          $ref: '#/components/schemas/Cursor'
...
    TournamentState:
      type: string
      title: TournamentState
      description: |
        Derived from the server clock on every read, never stored — so it
        cannot disagree with the schedule.
      enum: [SCHEDULED, OPEN, CLOSED]
      examples: ['OPEN']
    TournamentSummary:
      type: object
      title: TournamentSummary
      description: What a player needs in order to choose a tournament. Reveals no question content.
      required: [id, name, state, opensAt, closesAt, questions, maxAttempts]
      properties:
        id:
          $ref: '#/components/schemas/TournamentId'
        name:
          type: string
          maxLength: 200
        state:
          $ref: '#/components/schemas/TournamentState'
        opensAt:
          type: string
          format: date-time
          maxLength: 40
        closesAt:
          type: string
          format: date-time
          maxLength: 40
        questions:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
          description: How many questions an attempt will contain.
        timeLimitSeconds:
          type: [integer, 'null']
          format: int32
          minimum: 1
          maximum: 86400
          description: Per-attempt time limit. `null` means no limit.
        maxAttempts:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        bankId:
          allOf:
            - $ref: '#/components/schemas/QuestionBankId'
          description: |
            The bank this tournament draws from. Optional rather than required
            because it was added after the shape was published, and `/v1` is
            additive-only.
        scoringPolicy:
          $ref: '#/components/schemas/ScoringPolicy'
      examples:
        - id: trn_01a0136fb3937209a5719e8fb7db26bf
          name: Geography Weekly
          state: OPEN
          opensAt: '2026-08-18T09:00:00Z'
          closesAt: '2026-08-25T09:00:00Z'
          questions: 10
          timeLimitSeconds: 600
          maxAttempts: 1
...
    QuestionBankId:
      type: string
      title: QuestionBankId
      description: A question bank identifier. UUIDv7 rendered as 32 lowercase hex characters with a `bnk_` prefix.
      pattern: '^bnk_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['bnk_01a0136fb3937209a5719e8fb7db26bf']
...
    TournamentDraft:
      type: object
      title: TournamentDraft
      description: |
        Everything that defines a tournament. Used for both creation and
        amendment — `PATCH` replaces the whole draft rather than merging fields,
        so a client always sends the tournament it wants to exist.
      required: [name, bankId, opensAt, closesAt, questions, maxAttempts]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        bankId:
          $ref: '#/components/schemas/QuestionBankId'
        opensAt:
          type: string
          format: date-time
        closesAt:
          type: string
          format: date-time
          description: Must be after `opensAt`.
        questions:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
          description: |
            How many questions each attempt draws. Refused if the bank holds
            fewer current questions than this.
        timeLimitSeconds:
          type: [integer, 'null']
          format: int32
          minimum: 1
          maximum: 86400
          description: Per-attempt limit, or `null` for none.
        maxAttempts:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        scoringPolicy:
          $ref: '#/components/schemas/ScoringPolicy'
      examples:
        - name: Geography Weekly
          bankId: bnk_01a0136fb3937209a5719e8fb7db26bf
          opensAt: '2026-09-01T09:00:00Z'
          closesAt: '2026-09-08T09:00:00Z'
          questions: 10
          timeLimitSeconds: 600
          maxAttempts: 1
          scoringPolicy: BEST
    ScoringPolicy:
      type: string
      title: ScoringPolicy
      description: |
        Which attempt counts when a player has several. Applied when standings
        are **read**, not when a score is stored, so changing it takes effect
        immediately without recomputing history.

        Defaults to `FIRST` when omitted.
      enum: [BEST, FIRST, LAST, AVERAGE]
      default: FIRST
      examples: ['BEST']
...